Corrected most of Unit tests
