--- examples/todos-typescript/src/__tests__/components/app.test.js
@@ -1,24 +1,0 @@
-// @flow
-
-import React from 'react';
-import configureStore from 'redux-mock-store';
-import { shallow } from 'enzyme';
-
-import App from '../../components/App';
-
-const setup = (setupProps = {}) => {
-  const store = configureStore()();
-  const wrapper = shallow(<App store={store} />);
-
-  return {
-    store,
-    wrapper
-  };
-};
-
-describe('App', () => {
-  test('renders without crashing', () => {
-    const { wrapper } = setup();
-    expect(wrapper).toMatchSnapshot();
-  });
-});
