--- examples/todos-typescript/src/__tests__/components/footer.test.js
@@ -1,24 +1,0 @@
-// @flow
-
-import React from 'react';
-import configureStore from 'redux-mock-store';
-import { shallow } from 'enzyme';
-
-import Footer from '../../components/Footer';
-
-const setup = (setupProps = {}) => {
-  const store = configureStore()();
-  const wrapper = shallow(<Footer store={store} />);
-
-  return {
-    store,
-    wrapper
-  };
-};
-
-describe('Footer', () => {
-  test('renders without crashing', () => {
-    const { wrapper } = setup();
-    expect(wrapper).toMatchSnapshot();
-  });
-});
