--- examples/todos-typescript/src/__tests__/reducers/index.test.js
@@ -1,12 +1,0 @@
-// @flow
-
-import reducer from '../../reducers';
-
-describe('root reducer', () => {
-  it('should combine all reducers', () => {
-    expect(reducer({}, { type: '@@INIT' })).toEqual({
-      todos: [],
-      visibilityFilter: 'SHOW_ALL'
-    });
-  });
-});
