+++ examples/todos-typescript/src/__tests__/reducers/index.test.ts
@@ -1,0 +1,10 @@
+import reducer from '../../reducers';
+
+describe('root reducer', () => {
+  it('should combine all reducers', () => {
+    expect(reducer({}, { type: '@@INIT' })).toEqual({
+      todos: [],
+      visibilityFilter: 'SHOW_ALL'
+    });
+  });
+});
